+++ tests/conftest.py
@@ -16,6 +16,18 @@ from homematicip.connection import Connection
 from homematicip_demo.fake_cloud_server import AsyncFakeCloudServer
 from homematicip_demo.helper import *
 
+# content of conftest.py
+def pytest_configure(config):
+    import sys
+
+    sys._called_from_test = True
+
+
+def pytest_unconfigure(config):
+    import sys  # This was missing from the manual
+
+    del sys._called_from_test
+
 
 @pytest.fixture
 def ssl_ctx():